reduced accessTokenExpires with 10secs and replaced getSession with getServerSession as per next-auth recommendation
